util_lib: Add functionality to read elf notes
'vmcore-dmesg.c' already implements functionality to read
'vmcoreinfo' from vmcore file. The same can be used in other
features as well (one of which is described below), so there
is merit in moving this to the utility libraries (util_lib).

Since kernel version 4.19-rc5 (Commit 23c85094fe1895caefdd
["proc/kcore: add vmcoreinfo note to /proc/kcore"]), '/proc/kcore'
contains a new PT_NOTE which carries the VMCOREINFO information.

If the same is available, we can benefit by using it in 'kexec-tools'.
This is especially useful for architectures like arm64 as we can
get kernel symbols like 'PHYS_OFFSET' from the '/proc/kcore' itself
and use it to calculate 'phys_offset' before we make a call to
'set_phys_offset()'.

Subsequent patch(es) in this series will use the same feature
to read the 'vmcoreinfo' from '/proc/kcore'.

Signed-off-by: Bhupesh Sharma <bhsharma@redhat.com>